Finalize callback situation
begin_step and wait callbacks are needlessly complicated, and
controlling how often to step the MCU is really the responsibility of
the frontend. These two callbacks have been removed and the logic baked
into FE_RunInstance. The sample callback is fine since it provides a
zero-copy way to get sample data out of the emulator.